Sonar Rules Fixed

+++ MultiThread.Test/UnitTest1.cs
@@ -6,50 +6,7 @@ using Xunit;
 
 namespace MultiThread.Test
 {
-    //[TestClass]
-    //public class UnitTest1
-    //{
-    //    [TestMethod]
-    //    public void TestLoadList()
-    //    {
-    //        // LoadList metodunu çalýþtýr ve objeler listesini al
-    //        Program.LoadList();
-    //        List<object[]> objects = Program.objects;
-
-    //        // Test: objeler listesi doðru bir þekilde yüklenmiþ mi?
-    //        Assert.AreEqual(100, objects.Count);
-    //        for (int i = 0; i < 100; i++)
-    //        {
-    //            Assert.AreEqual(i, (int)objects[i][0]);
-    //            Assert.AreEqual($"Name{i}", (string)objects[i][1]);
-    //            Assert.AreEqual($"Surname{i}", (string)objects[i][2]);
-    //            Assert.AreEqual($"01/01/19{i % 100}", (string)objects[i][3]);
-    //            Assert.AreEqual("Status", (string)objects[i][4]);
-    //        }
-    //    }
-
-    //    [TestMethod]
-    //    public void TestCreatePdf()
-    //    {
-    //        // LoadList metodunu çalýþtýr ve objeler listesini al
-    //        Program.LoadList();
-    //        List<object[]> objects = Program.objects;
-
-    //        // PDF dosyalarýný oluþtur
-    //        Program.CreatePdf();
-
-    //        // Test: PDF dosyalarý doðru bir þekilde oluþturulmuþ mu?
-    //        foreach (object[] obj in objects)
-    //        {
-    //            int id = (int)obj[0];
-    //            string fileName = $"object_{id}.pdf";
-    //            string filePath = Path.Combine(Program.pdfPath, fileName);
-    //            Assert.IsTrue(File.Exists(filePath));
-    //        }
-    //    }
-    //}
-
-
+   
     [TestClass]
     public class UnitTest1
     {
@@ -104,7 +61,7 @@ namespace MultiThread.Test
             }
         }
     }
-
+     
     public class ProgramFixture : IDisposable
     {
         public List<object[]> objects { get; private set; }
@@ -114,8 +71,8 @@ namespace MultiThread.Test
         {
             // LoadList metodunu çalýþtýr ve objeler listesini al
             Program.LoadList();
-            objects = Program.objects;
-            pdfPath = Program.pdfPath;
+            objects = Program.Objects;
+            pdfPath = Program.PdfPath;
         }
 
         public void CreatePdf()
@@ -124,21 +81,37 @@ namespace MultiThread.Test
             Program.CreatePdf();
         }
 
-        public void Dispose()
+        protected virtual void Dispose(bool disposing)
         {
-            // PDF dosyalarýný sil
-            foreach (object[] obj in objects)
+            if (disposing)
             {
-                int id = (int)obj[0];
-                string fileName = $"object_{id}.pdf";
-                string filePath = Path.Combine(pdfPath, fileName);
-                if (File.Exists(filePath))
+                // managed kaynaklarý boþalt
+                foreach (object[] obj in objects)
                 {
-                    File.Delete(filePath);
+                    int id = (int)obj[0];
+                    string fileName = $"object_{id}.pdf";
+                    string filePath = Path.Combine(pdfPath, fileName);
+                    if (File.Exists(filePath))
+                    {
+                        File.Delete(filePath);
+                    }
                 }
             }
+
+            // unmanaged kaynaklarý boþalt
         }
 
+        ~ProgramFixture()
+        {
+            Dispose(false);
+        }
+
+        public void Dispose()
+        {
+            Dispose(true);
+            GC.SuppressFinalize(this);
+        }
     }
+
 }
